livestock lamb counter/lamb_counting/lamb_Counting/requirements.txt
inference==0.8.8
numpy  # Update this line with a compatible numpy version
onnxruntime
opencv-python
Pillow
streamlit
torch
transformers